[feat/ABOUT-31] Directory Structure 변경
