+++ app/src/main/java/com/w36495/about/domain/entity/Topic.kt
@@ -1,4 +1,4 @@
-package com.w36495.about.data
+package com.w36495.about.domain.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
